requirements.txt
flask==2.3.2
python-dotenv==1.0.0
protobuf==3.20.3
google-generativeai==0.3.2
gunicorn==20.1.0